# Conflicts:
#	gemrb/GUIScripts/GUILOAD.py
#	gemrb/GUIScripts/bg2/Game.py
#	gemrb/GUIScripts/bg2/Start.py
#	gemrb/GUIScripts/iwd2/GUILOAD.py
#	gemrb/core/GUI/Console.h
#	gemrb/core/ScriptedAnimation.cpp

+++ gemrb/core/PathFinder.h
@@ -65,11 +65,11 @@ enum {
 // to sort nodes by their (heuristic) distance from the destination
 class PQNode {
 public:
-	PQNode(Point p, unsigned int l) : point(p), dist(l) {};
+	PQNode(Point p, double l) : point(p), dist(l) {};
 	PQNode() : point(Point(0, 0)), dist(0) {};
 
 	Point point;
-	unsigned int dist;
+	double dist;
 
 	friend bool operator < (const PQNode &lhs, const PQNode &rhs) { return lhs.dist < rhs.dist;}
 	friend bool operator > (const PQNode &lhs, const PQNode &rhs){ return rhs < lhs; }
